Renaming NLP Services to NLP Providers (so that we can have a NLPService)
